advance 01

+++ src/controllers/cart.controllers.js
@@ -1,36 +1,111 @@
 const catchError = require('../utils/catchError');
 const Cart = require('../models/Cart');
+const Product = require('../models/Product');
+const Category = require('../models/Category');
+const ProductImg = require('../models/ProductImg');
 
-const getAll = catchError(async(req, res) => {
-    const results = await Cart.findAll();
+const getAll = catchError(async (req, res) => {
+
+    const userId = req.user.id
+
+    const results = await Cart.findAll({
+        where: { userId },
+        include: [
+            {
+                model: Product,
+                attributes: { exclude: ['createdAt', 'updatedAt'] },
+                include: [
+                    {
+                        model: Category,
+                        attributes: ['name']
+                    },
+                    {
+                        model: ProductImg
+                    }
+                ]
+            }
+        ]
+    });
     return res.json(results);
 });
 
-const create = catchError(async(req, res) => {
-    const result = await Cart.create(req.body);
+const create = catchError(async (req, res) => {
+
+    const { quantity, productId } = req.body
+    const userId = req.user.id
+
+
+    const product = await Product.findByPk(productId)
+
+    if (!product) {
+        return res.status(404).json({ message: 'Product not found' })
+    }
+
+    const body = {
+        userId, quantity, productId
+    }
+
+    const result = await Cart.create(body);
+
     return res.status(201).json(result);
 });
 
-const remove = catchError(async(req, res) => {
+const getOne = catchError(async (req, res) => {
+    
     const { id } = req.params;
-    const result = await Cart.destroy({ where: {id} });
-    if(!result) return res.sendStatus(404);
+    const userId = req.user.id
+
+    const result = await Cart.findByPk(id, {
+        where: { userId },
+        include: [
+            {
+                model: Product,
+                attributes: { exclude: ['createdAt', 'updatedAt'] },
+                include: [
+                    {
+                        model: Category,
+                        attributes: ['name']
+                    },
+                    {
+                        model: ProductImg
+                    }
+                ]
+            }
+        ]
+    });
+    if (!result) return res.sendStatus(404);
+    return res.json(result);
+});
+
+const remove = catchError(async (req, res) => {
+
+    const userId = req.user.id
+
+    const { id } = req.params;
+    const result = await Cart.destroy({ where: { id, userId } });
+    if (!result) return res.sendStatus(404);
     return res.sendStatus(204);
 });
 
-const update = catchError(async(req, res) => {
+const update = catchError(async (req, res) => {
+
     const { id } = req.params;
+    const userId = req.user.id
+    const { quantity } = req.body
+
+
     const result = await Cart.update(
-        req.body,
-        { where: {id}, returning: true }
+        { quantity },
+        { where: { id, userId }, returning: true }
     );
-    if(result[0] === 0) return res.sendStatus(404);
+    if (result[0] === 0) return res.sendStatus(404);
     return res.json(result[1][0]);
 });
 
 module.exports = {
     getAll,
     create,
+    getOne,
     remove,
     update
 }
